+++ src/main/webapp/app/components/layout-components/header.tsx
@@ -2,73 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { AppBar, Toolbar, IconButton, Typography, InputBase } from '@material-ui/core';
 // tslint:disable-next-line:no-submodule-imports
-import { fade } from '@material-ui/core/styles/colorManipulator';
-// tslint:disable-next-line:no-submodule-imports
 import { withStyles } from '@material-ui/core/styles';
 // tslint:disable-next-line:no-submodule-imports
 import MenuIcon from '@material-ui/icons/Menu';
 // tslint:disable-next-line:no-submodule-imports
 import SearchIcon from '@material-ui/icons/Search';
-
-const styles = theme => ({
-  root: {
-    width: '100%'
-  },
-  grow: {
-    flexGrow: 1
-  },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20
-  },
-  title: {
-    display: 'none',
-    [theme.breakpoints.up('sm')]: {
-      display: 'block'
-    }
-  },
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.25)
-    },
-    marginLeft: 0,
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      marginLeft: theme.spacing.unit,
-      width: 'auto'
-    }
-  },
-  searchIcon: {
-    width: theme.spacing.unit * 9,
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  inputRoot: {
-    color: 'inherit',
-    width: '100%'
-  },
-  inputInput: {
-    paddingTop: theme.spacing.unit,
-    paddingRight: theme.spacing.unit,
-    paddingBottom: theme.spacing.unit,
-    paddingLeft: theme.spacing.unit * 10,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: 120,
-      '&:focus': {
-        width: 200
-      }
-    }
-  }
-});
+import { headerSearchStyles } from 'app/components/search-header-jss';
 
 interface ISearchAppBarProps {
   classes: any;
@@ -108,4 +47,4 @@ class SearchAppBar extends React.Component<ISearchAppBarProps> {
   }
 }
 
-export const VibeHeader = withStyles(styles)(SearchAppBar);
+export const VibeHeader = withStyles(headerSearchStyles)(SearchAppBar);
