+++ src/main/webapp/app/components/layout-components/layout.scss
@@ -1,0 +1,3 @@
+[data-component='vibe-layout'] {
+  height: 100vh;
+}
